+++ src/azure-cli/azure/cli/command_modules/monitor/tests/latest/recordings/test_resource_location_arg_fmt.yaml
@@ -16,7 +16,7 @@ interactions:
     uri: https://management.azure.com/subscriptions/00000000-0000-0000-0000-000000000000/resourcegroups/test_resource_location_arg_fmt000001?api-version=2021-04-01
   response:
     body:
-      string: '{"id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test_resource_location_arg_fmt000001","name":"test_resource_location_arg_fmt000001","type":"Microsoft.Resources/resourceGroups","location":"eastus2","tags":{"product":"azurecli","cause":"automation","date":"2022-07-06T14:00:48Z"},"properties":{"provisioningState":"Succeeded"}}'
+      string: '{"id":"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test_resource_location_arg_fmt000001","name":"test_resource_location_arg_fmt000001","type":"Microsoft.Resources/resourceGroups","location":"eastus2","tags":{"product":"azurecli","cause":"automation","date":"2022-07-07T06:16:39Z"},"properties":{"provisioningState":"Succeeded"}}'
     headers:
       cache-control:
       - no-cache
@@ -25,7 +25,7 @@ interactions:
       content-type:
       - application/json; charset=utf-8
       date:
-      - Wed, 06 Jul 2022 14:00:53 GMT
+      - Thu, 07 Jul 2022 06:16:42 GMT
       expires:
       - '-1'
       pragma:
@@ -158,7 +158,7 @@ interactions:
       content-type:
       - application/json; charset=utf-8
       date:
-      - Wed, 06 Jul 2022 14:00:55 GMT
+      - Thu, 07 Jul 2022 06:16:44 GMT
       expires:
       - '-1'
       pragma:
@@ -291,7 +291,7 @@ interactions:
       content-type:
       - application/json; charset=utf-8
       date:
-      - Wed, 06 Jul 2022 14:00:57 GMT
+      - Thu, 07 Jul 2022 06:16:48 GMT
       expires:
       - '-1'
       pragma:
